requirements.txt
Pillow
numpy
black